refactor : 선물 테마 관련 코드 수정 - 피드백 반영 : useApiClient는 자체가 기능을 크게 억제함으로 각 컴포넌트의 훅 자체에서 useQuery를 사용하게 하고 그 대신 재사용성을 위해 공동으로 사용하는 queryFn 함수를 제네릭 타입을 이용해 정의
