+++ src/api/baseUrl.ts
@@ -1,0 +1,5 @@
+import axios from 'axios';
+
+export const instance = axios.create({
+  baseURL: import.meta.env.VITE_API_BASE_URL,
+});
