--- src/components/Home/PresentTheme/usePresentTheme.ts
@@ -1,48 +1,0 @@
-import { apiClient } from '@src/api/FetchData';
-import { useEffect, useState } from 'react';
-import type { Themes } from './Item/ThemeType';
-import { BASIC_ENDPOINT } from '@src/assets/endpoints';
-import type { HttpTypes } from '@src/api/HttpType';
-
-export const usePresentTheme = () => {
-  const [isLoading, setLoading] = useState<boolean>(true);
-  const [isError, setError] = useState<boolean>(false);
-  const [themes, setThemes] = useState<Themes>({
-    data: [
-      {
-        themeId: 0,
-        name: '',
-        image: 'none',
-      },
-    ],
-  });
-
-  useEffect(() => {
-    const fetchProductRanking = async () => {
-      try {
-        const apiRequestData = {
-          methods: 'GET' as HttpTypes,
-          requestName: BASIC_ENDPOINT.theme,
-          body: {},
-          params: '',
-          headers: null,
-        };
-        const data = await apiClient(apiRequestData);
-        setThemes(data);
-        setError(false);
-      } catch (error) {
-        console.error('Error fetching Product Ranking data:', error);
-        setError(true);
-      } finally {
-        setLoading(false);
-      }
-    };
-    fetchProductRanking();
-  }, []);
-
-  return {
-    themes,
-    isLoading,
-    isError,
-  };
-};
